InputStream, OutputStream: move types to a separate file
